requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://mirrors.cloud.tencent.com/pypi/simple
anyio==3.6.1; python_full_version >= '3.6.2'
certifi==2022.6.15; python_version >= '3.6'
click==8.1.3; python_version >= '3.7'
fastapi==0.78.0
gevent==21.12.0
greenlet==1.1.2; platform_python_implementation == 'CPython'
gunicorn==20.1.0
h11==0.12.0; python_version >= '3.6'
httpcore==0.15.0; python_version >= '3.7'
httptools==0.4.0
httpx==0.23.0
idna==3.3; python_version >= '3.5'
pycryptodome==3.15.0
pydantic==1.9.1; python_full_version >= '3.6.1'
python-dateutil==2.8.2
rfc3986[idna2008]==1.5.0
setuptools==63.1.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.2.0; python_version >= '3.5'
starlette==0.19.1; python_version >= '3.6'
typing-extensions==4.3.0; python_version >= '3.7'
uvicorn==0.18.2
uvloop==0.16.0
zope.event==4.5.0
zope.interface==5.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
